LayoutTests/inspector/styles-source-lines.html

<html>
<head>
<!--
    Line numbers are reported wrt the opened style tag baseline.
    Allow some space here to align local line numbers to be style + 10.
    Selector sourceLine is the line of the corresponding { following the
    selector text.
-->

<style type="text/css" media="screen">
body, .at_line_15::/*
Multiline comment
before pseudo element
              */before 
{
    color:red;
    content: "Before"
}
/*

Multiline comment followed with whitespace


*/

body, .at_line_29


{
    /* Comment in definition*/
    /* Comment in definition 2*/

    /* Comment in definition 3
    */font-family:/*  Comment in 
                      value  */courier;
    color blue;
}/*Multiline comment

followed by selector
*/body,
.at_line_43

{
    font-size:
        10px;
}


body, .at_line_50
{
    border: 1px solid
    red;
}
</style>

<script src="../http/tests/inspector/inspector-test.js"></script>
<script src="elements-tests.js"></script>
<script>

function doit()
{
    function callback(styles)
    {
        var matchedCSSRules = styles.matchedCSSRules;
        for (var i = 0; i < matchedCSSRules.length; ++i) {
            output("Line:" + (matchedCSSRules[i].sourceLine + 1) + "\t" + matchedCSSRules[i].selectorText);
        }
        notifyDone();
    }
    evaluateInWebInspector("frontend_dumpSourceLines", callback);
}


// Frontend functions.

function frontend_dumpSourceLines(testController)
{
    testController.waitUntilDone();
    frontend_expandDOMSubtreeAndRun(testController, null, frontend_getRawStyles);
}

function frontend_getRawStyles(testController)
{
    function callback(styles)
    {
        testController.notifyDone(styles);
    }
    InspectorBackend.getStyles(WebInspector.domAgent.document.body.id, true, callback);
}

</script>
</head>

<body onload="onload()">
<p>
Tests that proper source lines are reported for the parsed styles.
</p>

</body>
</html>
